backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.9
langchain==1.0.0
langchain-community==0.3.29
langchain-text-splitters==0.3.11
faiss-cpu==1.12.0
fastembed==0.4.2
pypdf==6.0.0
httpx==0.28.1
python-multipart==0.0.20
